address comments     - created permissions folder     - create viewmodel outside of permissions screen
